-Refactors repositories -Renames items in EventDetailViewModel.kt -Makes some if else statements when statements
