--- app/src/main/java/com/andrew/tasky/agenda/data/AgendaResponse.kt
@@ -1,7 +1,0 @@
-package com.andrew.tasky.agenda.data
-
-data class AgendaResponse(
-    val events: List<Event>,
-    val tasks: List<Task>,
-    val reminders: List<Reminder>
-)
