+++ app/src/main/java/com/andrew/tasky/agenda/data/agenda/AgendaItemDatabase.kt
@@ -1,6 +1,7 @@
-package com.andrew.tasky.agenda.domain.db
+package com.andrew.tasky.agenda.data.agenda
 
 import androidx.room.*
+import com.andrew.tasky.agenda.domain.db.Converters
 import com.andrew.tasky.agenda.domain.models.AgendaItem
 
 @Database(
