+++ app/src/main/java/com/andrew/tasky/agenda/data/event/EventRepositoryImpl.kt
@@ -1,0 +1,32 @@
+package com.andrew.tasky.agenda.data.event
+
+import com.andrew.tasky.agenda.domain.EventRepository
+import com.andrew.tasky.agenda.domain.models.AgendaItem
+import com.andrew.tasky.agenda.domain.models.Attendee
+import com.andrew.tasky.auth.data.AuthResult
+
+class EventRepositoryImpl() : EventRepository {
+    override suspend fun createEvent(): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun getEvent(eventId: String): AuthResult<AgendaItem> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun deleteEvent(eventId: String): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun updateEvent(): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun getAttendee(email: String): AuthResult<Attendee> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun deleteAttendee(eventId: String): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+}
