+++ app/src/main/java/com/andrew/tasky/agenda/data/task/TaskRepositoryImpl.kt
@@ -1,0 +1,23 @@
+package com.andrew.tasky.agenda.data.task
+
+import com.andrew.tasky.agenda.domain.TaskRepository
+import com.andrew.tasky.agenda.domain.models.AgendaItem
+import com.andrew.tasky.auth.data.AuthResult
+
+class TaskRepositoryImpl() : TaskRepository {
+    override suspend fun createTask(task: AgendaItem): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun updateTask(task: AgendaItem): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun getTask(taskId: String): AuthResult<AgendaItem> {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun deleteTask(taskId: String): AuthResult<Unit> {
+        TODO("Not yet implemented")
+    }
+}
